requirements.txt
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
pyjwt>=2.8.0
cryptography>=41.0.0
python-dotenv>=1.0.0
schedule>=1.2.0
yfinance>=0.2.0
